+++ Main.cs
@@ -58,6 +58,21 @@ public sealed partial class Main : BaseUnityPlugin
 
     void Awake()
     {
+//#if DEBUG
+//        foreach (var type in typeof(Game).Assembly.ExportedTypes.Where(x => !x.IsEnum))
+//        {
+//            foreach (var field in type.GetFields(BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Static)
+//                .Where(x => x.IsLiteral && !x.IsInitOnly))
+//            {
+//                var value = field.GetRawConstantValue();
+//                if (Equals(value, 3000f))
+//                    Logger.LogError($"{type.Name}.{field.Name}: {value} ({field.FieldType.Name})");
+//                else
+//                    Logger.LogWarning($"{type.Name}.{field.Name}: {value} ({field.FieldType.Name})");
+//            }
+//        }
+//#endif
+
         if (PluginVersion != PluginInformationalVersion)
             Logger.LogWarning($"You are running a pre-release version: {PluginInformationalVersion}");
         HarmonyInstance.PatchAll(Assembly.GetExecutingAssembly());
@@ -71,6 +86,7 @@ public sealed partial class Main : BaseUnityPlugin
 
     void Start()
     {
+
         StartCoroutine(CallExecute());
 
         IEnumerator<YieldInstruction> CallExecute()