+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### v0.2.16
-- New feature: remove mist from Mistlands
+- New features:
+    - Remove mist from Mistlands
+    - Auto name portals when automatically generated portal hub is enabled
 
 ### v0.2.15
 - Fix bug introduced in v0.2.14 which lead to many features not working correctly